:recycle: Move getAmountFeedbackAndError to utils
