+++ src/utils/validators.js
@@ -1,9 +1,10 @@
 import * as bitcoin from 'bitcoinjs-lib';
 import numeral from 'numeral';
-import getBtcConfig from './api/btc/config';
-import reg from './regex';
+import { fromRawLsk } from './lsk';
 import { tokenMap } from '../constants/tokens';
+import getBtcConfig from './api/btc/config';
 import i18n from '../i18n';
+import reg from './regex';
 
 /**
  * Validates the given address with respect to the tokenType
@@ -70,3 +71,20 @@ export const validateAmountFormat = ({
     message,
   };
 };
+
+export function getAmountFeedbackAndError({
+  value, fee, account, token,
+}) {
+  const { message, error } = validateAmountFormat({
+    value,
+    token,
+    locale: i18n.language,
+  });
+  let feedback = message;
+
+  const getMaxAmount = () => fromRawLsk(Math.max(0, account.balance - fee));
+  if (!error && parseFloat(getMaxAmount()) < numeral(value).value()) {
+    feedback = i18n.t('Provided amount is higher than your current balance.');
+  }
+  return { error, feedback };
+}
